Move exception mappers to separate package
