+++ src/main/java/com/github/eyefloaters/console/api/mapping/NotFoundExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.github.eyefloaters.console.api.support;
+package com.github.eyefloaters.console.api.mapping;
 
 import jakarta.ws.rs.NotFoundException;
 import jakarta.ws.rs.ext.Provider;
